Move filter class into transit.service
