+++ src/main/java/org/opentripplanner/apis/gtfs/datafetchers/RouteImpl.java
@@ -9,12 +9,12 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.opentripplanner.apis.gtfs.GraphQLRequestContext;
 import org.opentripplanner.apis.gtfs.GraphQLUtils;
-import org.opentripplanner.apis.gtfs.PatternByServiceDatesFilter;
 import org.opentripplanner.apis.gtfs.generated.GraphQLDataFetchers;
 import org.opentripplanner.apis.gtfs.generated.GraphQLTypes;
 import org.opentripplanner.apis.gtfs.generated.GraphQLTypes.GraphQLBikesAllowed;
 import org.opentripplanner.apis.gtfs.generated.GraphQLTypes.GraphQLTransitMode;
 import org.opentripplanner.apis.gtfs.mapping.BikesAllowedMapper;
+import org.opentripplanner.apis.gtfs.support.filter.PatternByDateFilterUtil;
 import org.opentripplanner.apis.gtfs.support.time.LocalDateRangeUtil;
 import org.opentripplanner.routing.alertpatch.EntitySelector;
 import org.opentripplanner.routing.alertpatch.TransitAlert;
@@ -183,7 +183,10 @@ public class RouteImpl implements GraphQLDataFetchers.GraphQLRoute {
       var args = new GraphQLTypes.GraphQLRoutePatternsArgs(environment.getArguments());
 
       if (LocalDateRangeUtil.hasServiceDateFilter(args.getGraphQLServiceDates())) {
-        var filter = new PatternByServiceDatesFilter(args.getGraphQLServiceDates(), transitService);
+        var filter = PatternByDateFilterUtil.ofGraphQL(
+          args.getGraphQLServiceDates(),
+          transitService
+        );
         return filter.filterPatterns(patterns);
       } else {
         return patterns;
